split MINIT for TextInfo
