+++ package.xml
@@ -87,6 +87,8 @@ Fix: ensure build directories are created for each source directory, fixing an i
             <file name="locale_options.h" role="src"/>
             <file name="locale_text_info.c" role="src"/>
             <file name="locale_text_info.h" role="src"/>
+            <file name="locale_text_info_arginfo.h" role="src"/>
+            <file name="locale_text_info.stub.php" role="src"/>
             <file name="locale_week_day.c" role="src"/>
             <file name="locale_week_day.h" role="src"/>
             <file name="locale_week_day_arginfo.h" role="src"/>
