+++ src/php/ecma_intl.c
@@ -52,7 +52,7 @@ PHP_MINIT_FUNCTION(ecma_intl) {
   registerEcmaIntlLocale();
   registerEcmaIntlLocaleCharacterDirection();
   registerEcmaIntlLocaleOptions();
-  registerEcmaIntlLocaleTextInfo();
+  PHP_MINIT(ecma_intl_locale_textinfo)(INIT_FUNC_ARGS_PASSTHRU);
   PHP_MINIT(ecma_intl_locale_weekday)(INIT_FUNC_ARGS_PASSTHRU);
   PHP_MINIT(ecma_intl_locale_weekinfo)(INIT_FUNC_ARGS_PASSTHRU);
